+++ KING_Event_Team/KING_Event_Team/src/main/java/br/com/fatec/controller/GerenEventosController.java
@@ -16,6 +16,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Optional;
 import java.util.ResourceBundle;
 import javafx.beans.property.ReadOnlyStringWrapper;
 import javafx.collections.FXCollections;
@@ -25,6 +26,7 @@ import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.Label;
 import javafx.scene.control.TableColumn;
@@ -139,19 +141,25 @@ public class GerenEventosController implements Initializable{
             App.mensagem("AVISO", "Selecione um Evento!", Alert.AlertType.WARNING);
             return;
         }
+        Alert alerta = new Alert(Alert.AlertType.CONFIRMATION);
+        alerta.setTitle("CONFIRMAÇÃO");
+        alerta.setHeaderText("Deseja excluir este item?");
+        Optional<ButtonType> resultado = alerta.showAndWait();
+        if(resultado.isPresent() && resultado.get() == ButtonType.OK){
         
-        try
-        {
-            EventoDAO dao = new EventoDAO();
-            dao.remover(evento);
-            App.mensagem("APAGADO", "Evento apagado com sucesso!");
+            try
+            {
+                EventoDAO dao = new EventoDAO();
+                dao.remover(evento);
+                App.mensagem("APAGADO", "Evento apagado com sucesso!");
+            }
+            catch(SQLException ex)
+            {
+                App.mensagem("ERRO", "Erro ao Deletar", Alert.AlertType.ERROR);
+            }
+
+            preencherTabela();
         }
-        catch(SQLException ex)
-        {
-            App.mensagem("ERRO", "Erro ao Deletar", Alert.AlertType.ERROR);
-        }
-        
-        preencherTabela();
     }
 
     @FXML
